Moved geocoding to com.graphhopper.geocoding.client
