+++ geocoding-config.json
@@ -1,9 +1,9 @@
 {
-  "apiPackage": "com.graphhopper.geocoding-client.api",
-  "modelPackage": "com.graphhopper.geocoding-client.model",
-  "invokerPackage": "com.graphhopper.geocoding-client",
-  "groupId":"com.graphhopper",
-  "artifactVersion":"0.9-SNAPSHOT",
+  "apiPackage": "com.graphhopper.geocoding.client.api",
+  "modelPackage": "com.graphhopper.geocoding.client.model",
+  "invokerPackage": "com.graphhopper.geocoding.client",
+  "groupId": "com.graphhopper",
+  "artifactVersion": "0.9-SNAPSHOT",
   "library": "okhttp-gson",
   "artifactId": "directions-api-java-client-geocoding"
 }
