+++ geocoding/build.gradle
@@ -2,7 +2,7 @@ apply plugin: 'idea'
 apply plugin: 'eclipse'
 
 group = 'com.graphhopper'
-version = '0.8-SNAPSHOT'
+version = '0.9-SNAPSHOT'
 
 buildscript {
     repositories {
