+++ geocoding/build.sbt
@@ -2,7 +2,7 @@ lazy val root = (project in file(".")).
   settings(
     organization := "com.graphhopper",
     name := "directions-api-java-client-geocoding",
-    version := "0.8-SNAPSHOT",
+    version := "0.9-SNAPSHOT",
     scalaVersion := "2.11.4",
     scalacOptions ++= Seq("-feature"),
     javacOptions in compile ++= Seq("-Xlint:deprecation"),
