+++ geocoding/src/main/AndroidManifest.xml
@@ -1,3 +1,3 @@
-<manifest package="com.graphhopper.geocoding_client" xmlns:android="http://schemas.android.com/apk/res/android">
+<manifest package="com.graphhopper.geocoding.client" xmlns:android="http://schemas.android.com/apk/res/android">
     <application />
 </manifest>
