+++ geocoding/src/main/java/com/graphhopper/geocoding/client/JSON.java
@@ -23,7 +23,7 @@
  */
 
 
-package com.graphhopper.geocoding_client;
+package com.graphhopper.geocoding.client;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
